Postpone fetching off aggregate events so that we don't need blocking code.
