+++ Source/Aggregates/AggregateOf.cs
@@ -1,10 +1,6 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-using System;
-using System.Linq;
-using System.Reflection;
-using System.Threading;
 using Dolittle.SDK.Events;
 using Dolittle.SDK.Events.Store;
 using Microsoft.Extensions.Logging;
@@ -43,111 +39,10 @@ namespace Dolittle.SDK.Aggregates
 
         /// <inheritdoc/>
         public IAggregateRootOperations<TAggregateRoot> Get(EventSourceId eventSourceId)
-        {
-            if (TryGetAggregateRoot(eventSourceId, out var aggregateRoot, out var exception))
-            {
-                ReApplyEvents(aggregateRoot);
-                return new AggregateRootOperations<TAggregateRoot>(
-                    _eventStore,
-                    aggregateRoot,
-                    _eventTypes,
-                    _loggerFactory.CreateLogger<AggregateRootOperations<TAggregateRoot>>());
-            }
-
-            throw new CouldNotGetAggregateRoot(typeof(TAggregateRoot), eventSourceId, exception.Message);
-        }
-
-        bool TryGetAggregateRoot(EventSourceId eventSourceId, out TAggregateRoot aggregateRoot, out Exception exception)
-        {
-            try
-            {
-                exception = default;
-                _logger.LogDebug(
-                    "Getting aggregate root {AggregateRoot} with event source id {EventSource}",
-                    typeof(TAggregateRoot),
-                    eventSourceId);
-                aggregateRoot = CreateAggregateRoot(eventSourceId);
-                return true;
-            }
-            catch (Exception ex)
-            {
-                aggregateRoot = default;
-                exception = ex;
-                return false;
-            }
-        }
-
-        void ReApplyEvents(TAggregateRoot aggregateRoot)
-        {
-            var eventSourceId = aggregateRoot.EventSourceId;
-            var aggregateRootId = aggregateRoot.GetAggregateRootId();
-            _logger.LogDebug(
-                "Re-applying events for {AggregateRoot} with aggregate root id {AggregateRootId} with event source id {EventSourceId}",
-                typeof(TAggregateRoot),
-                aggregateRootId,
-                eventSourceId);
-
-            var committedEvents = _eventStore.FetchForAggregate(aggregateRootId, eventSourceId, CancellationToken.None).GetAwaiter().GetResult();
-            if (committedEvents.HasEvents)
-            {
-                _logger.LogTrace("Re-applying {NumberOfEvents} events", committedEvents.Count);
-                aggregateRoot.ReApply(committedEvents);
-            }
-            else
-            {
-                _logger.LogTrace("No events to re-apply");
-            }
-        }
-
-        TAggregateRoot CreateAggregateRoot(EventSourceId eventSourceId)
-        {
-            var aggregateRootType = typeof(TAggregateRoot);
-            ThrowIfInvalidConstructor(aggregateRootType);
-            var constructor = typeof(TAggregateRoot).GetConstructors().Single();
-
-            var aggregateRoot = GetInstanceFrom(eventSourceId, constructor);
-            ThrowIfCouldNotCreateAggregateRoot(aggregateRoot);
-            return aggregateRoot;
-        }
-
-        TAggregateRoot GetInstanceFrom(EventSourceId id, ConstructorInfo constructor)
-            => constructor.Invoke(
-                new object[]
-                {
-                    id
-                }) as TAggregateRoot;
-
-        void ThrowIfInvalidConstructor(Type type)
-        {
-            ThrowIfNotOneConstructor(type);
-            ThrowIfConstructorIsInvalid(type, type.GetConstructors().Single());
-        }
-
-        void ThrowIfNotOneConstructor(Type type)
-        {
-            if (type.GetConstructors().Length != 1)
-                throw new InvalidAggregateRootConstructorSignature(type, "expected only a single constructor");
-        }
-
-        void ThrowIfConstructorIsInvalid(Type type, ConstructorInfo constructor)
-        {
-            var parameters = constructor.GetParameters();
-            ThrowIfIncorrectParameter(type, parameters);
-        }
-
-        void ThrowIfIncorrectParameter(Type type, ParameterInfo[] parameters)
-        {
-            if (parameters.Length != 1 ||
-                (parameters[0].ParameterType != typeof(Guid) &&
-                parameters[0].ParameterType != typeof(EventSourceId)))
-            {
-                throw new InvalidAggregateRootConstructorSignature(type, $"expected only one parameter and it must be of type {typeof(Guid)} or {typeof(EventSourceId)}");
-            }
-        }
-
-        void ThrowIfCouldNotCreateAggregateRoot(TAggregateRoot aggregateRoot)
-        {
-            if (aggregateRoot == default) throw new CouldNotCreateAggregateRootInstance(typeof(TAggregateRoot));
-        }
+            => new AggregateRootOperations<TAggregateRoot>(
+                eventSourceId,
+                _eventStore,
+                _eventTypes,
+                _loggerFactory.CreateLogger<AggregateRootOperations<TAggregateRoot>>());
     }
 }
